requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.1
fake-useragent>=1.4.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

